Move plCmdParser to CoreLib.
