+++ Sources/Tests/NucleusTests/CMakeLists.txt
@@ -1,2 +1,1 @@
 add_subdirectory(pnEncryptionTest)
-add_subdirectory(pnUtilsTest)
